refactor: organize state machine into a dedicated package
- Created a new `states/` package to contain all state-related classes
- Moved BaseState, TitleState, PlayState, and PauseState into the package
- Updated all import statements to use the new package structure
- Added __init__.py with explicit exports for clean importing
- Verified all state transitions and functionality remain intact

This improves code organization and provides a clear structure
for adding future game states like GameOverState and SettingsState.
